+++ packages/elf-service/src/search/filter.rs
@@ -3,6 +3,6 @@ mod impact;
 mod parser;
 mod value;
 
-pub(crate) use self::{impact::SearchFilterImpact, parser::SearchFilter};
+pub(crate) use self::{impact::SearchFilterImpact, parser::search_filter::SearchFilter};
 
 #[cfg(test)] mod tests;
