+++ packages/elf-service/src/search/filter/parser/constants.rs
@@ -1,0 +1,5 @@
+pub(in crate::search::filter) const SEARCH_FILTER_EXPR_SCHEMA_V1: &str = "search_filter_expr/v1";
+pub(in crate::search::filter) const MAX_FILTER_DEPTH: usize = 8;
+pub(in crate::search::filter) const MAX_FILTER_NODES: usize = 128;
+pub(in crate::search::filter) const MAX_IN_LIST_ITEMS: usize = 128;
+pub(in crate::search::filter) const MAX_STRING_BYTES: usize = 512;
